+++ astropy/wcs/wcsapi/high_level_api.py
@@ -130,7 +130,7 @@ class HighLevelWCSMixin(BaseHighLevelWCS):
         unique_match = True
         for w in world_objects:
             matches = []
-            for key, (klass, _, _) in classes.items():
+            for key, (klass, *_) in classes.items():
                 if isinstance(w, klass):
                     matches.append(key)
             if len(matches) == 1:
@@ -147,7 +147,14 @@ class HighLevelWCSMixin(BaseHighLevelWCS):
 
         if unique_match:
 
-            for key, (klass, args, kwargs) in classes.items():
+            for key, (klass, args, kwargs, *rest) in classes.items():
+
+                if len(rest) == 0:
+                    klass_gen = klass
+                elif len(rest) == 1:
+                    klass_gen = rest[0]
+                else:
+                    raise ValueError("Tuples in world_axis_object_classes should have length 3 or 4")
 
                 # FIXME: For now SkyCoord won't auto-convert upon initialization
                 # https://github.com/astropy/astropy/issues/7689
@@ -158,12 +165,21 @@ class HighLevelWCSMixin(BaseHighLevelWCS):
                     else:
                         objects[key] = world_by_key[key]
                 else:
-                    objects[key] = klass(world_by_key[key], *args, **kwargs)
+                    objects[key] = klass_gen(world_by_key[key], *args, **kwargs)
 
         else:
 
             for ikey, key in enumerate(classes):
-                klass, args, kwargs = classes[key]
+
+                klass, args, kwargs, *rest = classes[key]
+
+                if len(rest) == 0:
+                    klass_gen = klass
+                elif len(rest) == 1:
+                    klass_gen = rest[0]
+                else:
+                    raise ValueError("Tuples in world_axis_object_classes should have length 3 or 4")
+
                 w = world_objects[ikey]
                 if not isinstance(w, klass):
                     raise ValueError("Expected the following order of world "
@@ -178,12 +194,15 @@ class HighLevelWCSMixin(BaseHighLevelWCS):
                     else:
                         objects[key] = w
                 else:
-                    objects[key] = klass(w, *args, **kwargs)
+                    objects[key] = klass_gen(w, *args, **kwargs)
 
         # We now extract the attributes needed for the world values
         world = []
         for key, _, attr in components:
-            world.append(rec_getattr(objects[key], attr))
+            if callable(attr):
+                world.append(attr(objects[key]))
+            else:
+                world.append(rec_getattr(objects[key], attr))
 
         # Finally we convert to pixel coordinates
         pixel = self.low_level_wcs.world_to_pixel_values(*world)
@@ -223,8 +242,14 @@ class HighLevelWCSMixin(BaseHighLevelWCS):
         result = []
 
         for key in default_order(components):
-            klass, ar, kw = classes[key]
-            result.append(klass(*args[key], *ar, **kwargs[key], **kw))
+            klass, ar, kw, *rest = classes[key]
+            if len(rest) == 0:
+                klass_gen = klass
+            elif len(rest) == 1:
+                klass_gen = rest[0]
+            else:
+                raise ValueError("Tuples in world_axis_object_classes should have length 3 or 4")
+            result.append(klass_gen(*args[key], *ar, **kwargs[key], **kw))
 
         if len(result) == 1:
             return result[0]