+++ CHANGES.rst
@@ -264,11 +264,13 @@ astropy.wcs
   slicing operations in turn. [#9210]
 
 - Added a ``wcs_info_str`` function to ``astropy.wcs.wcsapi`` to show a summary
-- Added a ``wcs_as_str`` function to ``astropy.wcs.wcsapi`` to show a summary
-- Added a ``wcs_info_str`` function to ``astropy.wcs.wcsapi`` to show a summary
   of an APE-14-compliant WCS as a string. [#8546, #9207]
+
 - Added two new optional attributes to the APE 14 low-level WCS: ``pixel_axis_names``
   and ``world_axis_names``. [#9156]
+
+- Updated the WCS class to now correctly take and return ``Time`` objects in the
+  high-level APE 14 API (e.g. ``pixel_to_world``. [#9376]
 
 API Changes
 -----------